Adds fallback mechanism for credential branding
