+++ core/App/screens/CredentialDetails.tsx
@@ -44,7 +44,7 @@ const CredentialDetails: React.FC<CredentialDetailsProps> = ({ navigation, route
   const credential = useCredentialById(credentialId)
   const credentialConnectionLabel = getCredentialConnectionLabel(credential)
   const { TextTheme, ColorPallet } = useTheme()
-  const { OCABundle, record } = useConfiguration()
+  const { OCABundleResolver, record } = useConfiguration()
 
   useEffect(() => {
     if (!agent) {
@@ -95,7 +95,7 @@ const CredentialDetails: React.FC<CredentialDetailsProps> = ({ navigation, route
       const date = new Date(credential.revocationNotification.revocationDate)
       setRevocationDate(date.toLocaleDateString(i18n.language, dateFormatOptions))
     }
-    OCABundle.presentationFields(credential as CredentialExchangeRecord, i18n.language).then((fields) =>
+    OCABundleResolver.presentationFields(credential as CredentialExchangeRecord, i18n.language).then((fields) =>
       setFields(fields)
     )
   }, [credential])
